+++ src/main/java/ewha/capston/cockChat/domain/chat/service/ChatService.java
@@ -40,53 +40,6 @@ public class ChatService {
         mongoChatRepository.save(chat);
     }
 
-    /* 체팅방 생성 */
-    public ResponseEntity<ChatRoomResponseDto> createChatRoom(ChatRoomRequestDto requestDto, Member member) {
-
-        /* 6자리의 랜덤 문자열 생성 : 중복 없도록 생성 */
-        String identifier = null;
-        while(Boolean.TRUE){
-            identifier = generateRandomMixStr(6,true);
-            if(chatRoomRepository.existsByIdentifier(identifier) == Boolean.FALSE) break;
-        }
-
-        /* 채팅방 생성 */
-        ChatRoom chatRoom = chatRoomRepository.save(ChatRoom.builder()
-                .roomName(requestDto.getRoomName())
-                .roomType(requestDto.getRoomType())
-                .nicknameType(requestDto.getNicknameType())
-                .password(requestDto.getPassword())
-                .identifier(identifier)
-                .chatRoomImgUrl(requestDto.getChatRoomImgUrl())
-                .build());
-
-
-        /* 채팅방 생성인을 participant 로 채팅방에 참가. <- 이 부분 우선 주석 처리 */
-        /*
-        String roomNickname = "별명";
-        if(chatRoom.getNicknameType() == Boolean.FALSE) roomNickname = ; // 별명 사용 채팅방인 경우
-        else roomNickname = member.getNickname(); // 실명 사용 채팅방인 경우
-        Participant participant = participantRepository.save(Participant.builder()
-                        .chatRoom(chatRoom)
-                        .member(member)
-                        .isOwner(Boolean.TRUE)
-                        .roomNickname(roomNickname)
-                        .participantImgUrl(member.getProfileImgUrl())
-                        .build()
-        );
-         */
-
-        return ResponseEntity.status(HttpStatus.CREATED)
-                .body(ChatRoomResponseDto.of(chatRoom));
-    }
-
-    /* 채팅방 고유 코드로 채팅방 조회 */
-    public ResponseEntity<ChatRoomResponseDto> getChatRoomByIdentifier(Member member, String code) {
-        ChatRoom chatRoom = chatRoomRepository.findByIdentifier(code)
-                .orElseThrow(()->new CustomException(ErrorCode.INVALID_ROOM));
-        return ResponseEntity.status(HttpStatus.OK)
-                .body(ChatRoomResponseDto.of(chatRoom));
-    }
 
     /* 메시지 보내기 */
     public void sendMessage(Long roomId, ChatMessageRequestDto requestDto) {
@@ -115,19 +68,4 @@ public class ChatService {
         messagingTemplate.convertAndSend("/topic/public/" + roomId, chat);
     }
 
-
-
-    /* 랜덤 문자열 생성 */
-    public  String generateRandomMixStr(int length, boolean isUpperCase) {
-        String characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
-
-        SecureRandom random = new SecureRandom();
-        StringBuilder sb = new StringBuilder(length);
-
-        for (int i = 0; i < length; i++) {
-            int index = random.nextInt(characters.length());
-            sb.append(characters.charAt(index));
-        }
-        return isUpperCase ? sb.toString() : sb.toString().toLowerCase();
-    }
 }
